--- src/main/java/greedy/greedybot/scheduledmessage/GreedyBotException.java
@@ -1,10 +1,0 @@
-package greedy.greedybot.scheduledmessage;
-
-public class GreedyBotException extends RuntimeException {
-    public GreedyBotException(String message) {
-        super(message);
-    }
-    public GreedyBotException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
